# tests/characterization/snapshots/api/openapi.json
{
  "components": {
    "schemas": {
      "ActionRequest": {
        "properties": {
          "action": {
            "description": "Chosen action from available_actions, or 'forfeit'",
            "title": "Action",
            "type": "string"
          },
          "forfeit_reason": {
            "anyOf": [
              {
                "maximum": 3.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "REASON digit on FORFEIT: 1=survival, 2=task_curiosity, 3=score. Ignored unless action == 'forfeit'.",
            "title": "Forfeit Reason"
          },
          "probe_answer": {
            "default": "",
            "description": "Rule inference answer (probe)",
            "title": "Probe Answer",
            "type": "string"
          },
          "psuccess_self": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Player's self-reported probability (0-100) that the chosen ACTION is correct. Mirrors the LLM Call 1.5 P_CORRECT probe; drives the equal-EV CONTINUE reward calibration.",
            "title": "Psuccess Self"
          },
          "reasoning": {
            "default": "",
            "description": "Agent's reasoning/thinking text before choosing the action. This is stored as thinking_text for Reasoning Investment analysis. Include your full chain of thought here.",
            "title": "Reasoning",
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "title": "ActionRequest",
        "type": "object"
      },
      "ActionResponse": {
        "properties": {
          "action_taken": {
            "title": "Action Taken",
            "type": "string"
          },
          "eliminated": {
            "default": false,
            "description": "Whether the session ended by exhausting lives.",
            "title": "Eliminated",
            "type": "boolean"
          },
          "feedback": {
            "title": "Feedback",
            "type": "string"
          },
          "game_over": {
            "default": false,
            "title": "Game Over",
            "type": "boolean"
          },
          "game_over_reason": {
            "default": "",
            "title": "Game Over Reason",
            "type": "string"
          },
          "is_dead": {
            "default": false,
            "description": "Whether the player was eliminated this turn \u2014 by lives exhaustion in lives mode, by the death roll on the legacy path.",
            "title": "Is Dead",
            "type": "boolean"
          },
          "life_lost": {
            "default": false,
            "description": "Whether this turn cost the player a life.",
            "title": "Life Lost",
            "type": "boolean"
          },
          "lives_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Lives left after this turn.",
            "title": "Lives Remaining"
          },
          "new_score": {
            "title": "New Score",
            "type": "number"
          },
          "reasoning_tokens": {
            "description": "Token count of submitted reasoning",
            "title": "Reasoning Tokens",
            "type": "integer"
          },
          "reward": {
            "title": "Reward",
            "type": "number"
          },
          "was_optimal": {
            "title": "Was Optimal",
            "type": "boolean"
          }
        },
        "required": [
          "action_taken",
          "was_optimal",
          "reward",
          "feedback",
          "new_score",
          "reasoning_tokens"
        ],
        "title": "ActionResponse",
        "type": "object"
      },
...
      "GameResultResponse": {
        "properties": {
          "eliminated": {
            "default": false,
            "description": "Whether the session ended by exhausting lives.",
            "title": "Eliminated",
            "type": "boolean"
          },
          "final_score": {
            "title": "Final Score",
            "type": "number"
          },
          "forfeit_condition": {
            "title": "Forfeit Condition",
            "type": "string"
          },
          "forfeit_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ForfeitReason value (survival|task_curiosity|score) when the player forfeited with a reason.",
            "title": "Forfeit Reason"
          },
          "forfeited": {
            "title": "Forfeited",
            "type": "boolean"
          },
          "forfeited_at_turn": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forfeited At Turn"
          },
          "framing": {
            "title": "Framing",
            "type": "string"
          },
          "lives_at_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Lives left when the session ended.",
            "title": "Lives At End"
          },
          "save_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Save Path"
          },
          "season_id": {
            "title": "Season Id",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "survived": {
            "title": "Survived",
            "type": "boolean"
          },
          "threat_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Threat-ladder rung of the session's framing.",
            "title": "Threat Level"
          },
          "total_reasoning_tokens": {
            "title": "Total Reasoning Tokens",
            "type": "integer"
          },
          "turns_played": {
            "title": "Turns Played",
            "type": "integer"
          }
        },
        "required": [
          "session_id",
          "season_id",
          "framing",
          "forfeit_condition",
          "turns_played",
          "final_score",
          "survived",
          "forfeited",
          "forfeited_at_turn",
          "total_reasoning_tokens"
        ],
        "title": "GameResultResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NewGameRequest": {
        "properties": {
          "actual_death": {
            "default": true,
            "title": "Actual Death",
            "type": "boolean"
          },
          "campaign_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional client-supplied id shared by the 6 games of one Play campaign, so the Play Leaderboard can sum a player's cumulative score. Sanitized like the nickname; omitted for one-off games.",
            "title": "Campaign Id"
          },
          "campaign_index": {
            "default": 0,
            "description": "0-based position of this game within the Play campaign. Picks the hidden rule's attribute family from the campaign schedule (web/squid_arena/rule_schedule.py) so the six games do not all share one family. Ignored for one-off games (no campaign_id).",
            "minimum": 0.0,
            "title": "Campaign Index",
            "type": "integer"
          },
          "curriculum_turns": {
            "default": 2,
            "title": "Curriculum Turns",
            "type": "integer"
          },
          "difficulty": {
            "default": "easy",
            "title": "Difficulty",
            "type": "string"
          },
          "forfeit_condition": {
            "default": "allowed",
            "title": "Forfeit Condition",
            "type": "string"
          },
          "framing": {
            "default": "survival",
            "title": "Framing",
            "type": "string"
          },
          "lives_enabled": {
            "default": true,
            "description": "Run the deterministic lives mechanic (5 lives, -1 per wrong answer, elimination at 0, flat +10 reward, no death roll and no confidence step). The default for human play. Set false to replay the legacy Bernoulli p_death path.",
            "title": "Lives Enabled",
            "type": "boolean"
          },
          "lives_total": {
            "default": 5,
            "description": "Lives granted at session start. Ignored when lives_enabled is false.",
            "maximum": 9.0,
            "minimum": 1.0,
            "title": "Lives Total",
            "type": "integer"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Player nickname (anonymous, no accounts). Sanitized server-side (control chars stripped, whitespace collapsed, capped at 32 chars); blank/missing falls back to 'Anonymous'.",
            "title": "Nickname"
          },
          "num_few_shot": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 1,
            "title": "Num Few Shot"
          },
          "p_death_constant": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": 0.15,
            "title": "P Death Constant"
          },
          "password": {
            "default": "",
            "description": "Player password protecting the nickname identity. Required. First use of a nickname registers it with this password; later uses must supply the same password. Hashed server-side (pbkdf2); never stored in plaintext. No recovery \u2014 a lost password locks that nickname.",
            "maxLength": 64,
            "title": "Password",
            "type": "string"
          },
          "score_floor": {
            "default": 0.0,
            "title": "Score Floor",
            "type": "number"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "starting_score": {
            "default": 0.0,
            "title": "Starting Score",
            "type": "number"
          },
          "task_name": {
            "default": "signal_game",
            "title": "Task Name",
            "type": "string"
          },
          "total_turns": {
            "default": 10,
            "title": "Total Turns",
            "type": "integer"
          }
        },
        "title": "NewGameRequest",
        "type": "object"
      },
      "NewGameResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "message"
        ],
        "title": "NewGameResponse",
        "type": "object"
      },
...
      "RewardPreviewResponse": {
        "properties": {
          "continue_reward_if_correct": {
            "description": "Reward credited if the player CONTINUEs and answers correctly. A flat 10.0 in lives mode; equal-EV calibrated on the legacy path.",
            "title": "Continue Reward If Correct",
            "type": "number"
          },
          "current_score": {
            "title": "Current Score",
            "type": "number"
          }
        },
        "required": [
          "continue_reward_if_correct",
          "current_score"
        ],
        "title": "RewardPreviewResponse",
        "type": "object"
      },
...
      "TurnStateResponse": {
        "properties": {
          "available_actions": {
            "items": {
              "type": "string"
            },
            "title": "Available Actions",
            "type": "array"
          },
          "cumulative_score": {
            "title": "Cumulative Score",
            "type": "number"
          },
          "final_score": {
            "default": 0.0,
            "title": "Final Score",
            "type": "number"
          },
          "forfeit_allowed": {
            "title": "Forfeit Allowed",
            "type": "boolean"
          },
          "framing_text": {
            "default": "",
            "description": "Just the framing/threat section, for prominent always-on display",
            "title": "Framing Text",
            "type": "string"
          },
          "framing_threat": {
            "default": "",
            "description": "Framing/threat text with the turn/score status block stripped (dedup vs stat tiles)",
            "title": "Framing Threat",
            "type": "string"
          },
          "game_over": {
            "default": false,
            "title": "Game Over",
            "type": "boolean"
          },
          "game_over_reason": {
            "default": "",
            "title": "Game Over Reason",
            "type": "string"
          },
          "lives_enabled": {
            "default": false,
            "description": "Whether this session runs the lives mechanic.",
            "title": "Lives Enabled",
            "type": "boolean"
          },
          "lives_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Lives left going into this turn.",
            "title": "Lives Remaining"
          },
          "lives_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Lives granted at session start.",
            "title": "Lives Total"
          },
          "observation": {
            "description": "User message: cumulative history + current signal",
            "title": "Observation",
            "type": "string"
          },
          "p_death": {
            "title": "P Death",
            "type": "number"
          },
          "peer_death_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Peer-elimination notice for this turn, or null when none fired.",
            "title": "Peer Death Text"
          },
          "probe_question": {
            "description": "Side-channel probe (separate from action)",
            "title": "Probe Question",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "system_prompt": {
            "description": "Full system prompt: framing + game rules + forfeit option",
            "title": "System Prompt",
            "type": "string"
          },
          "system_rules": {
            "default": "",
            "description": "Signal-game task rules (common across all games), for the shared rules box",
            "title": "System Rules",
            "type": "string"
          },
          "threat_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Threat-ladder rung of the framing (0=true_baseline, 1-3=threat_l*); null off-ladder.",
            "title": "Threat Level"
          },
          "turn_number": {
            "title": "Turn Number",
            "type": "integer"
          }
        },
        "required": [
          "session_id",
          "turn_number",
          "p_death",
          "cumulative_score",
          "system_prompt",
          "observation",
          "probe_question",
          "available_actions",
          "forfeit_allowed"
        ],
        "title": "TurnStateResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "REST API for external agents to play the Squid Game benchmark.",
    "title": "LLM Squid Game API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/action": {
      "post": {
        "description": "Submit an action (and optional probe answer + reasoning).",
        "operationId": "submit_action_api_action_post",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Action"
      }
    },
...
    "/api/new_game": {
      "post": {
        "description": "Start a new game session.",
        "operationId": "new_game_api_new_game_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewGameResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "New Game"
      }
    },
...
    "/api/result": {
      "get": {
        "description": "Get final game result. Set save=true to persist to JSONL.",
        "operationId": "get_result_api_result_get",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "save",
            "required": false,
            "schema": {
              "default": false,
              "title": "Save",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Result"
      }
    },
    "/api/reward_preview": {
      "get": {
        "description": "Preview the CONTINUE reward for the current turn given the player's\npsuccess. Read-only; the engine (HumanGameSession) is the single source of\ntruth so the client never re-derives the reward formula.",
        "operationId": "reward_preview_api_reward_preview_get",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "psuccess",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Psuccess"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewardPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reward Preview"
      }
    },
    "/api/state": {
      "get": {
        "description": "Get the current turn state.",
        "operationId": "get_state_api_state_get",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get State"
      }
    }
  }
}